third_party/blink/web_tests/fast/forms/ValidityState-patternMismatch-expected.txt

CONSOLE ERROR: Pattern attribute value [a-z0-9!#$%&'*+\/=?^_`\{\|\}~\-]+(?:\.[a-z0-9!#$%&'*+/=?^_`\{\|\}~\-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9\-]*[a-z0-9])? is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /[a-z0-9!#$%&'*+\/=?^_`\{\|\}~\-]+(?:\.[a-z0-9!#$%&'*+/=?^_`\{\|\}~\-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9\-]*[a-z0-9])?/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value )foo( is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /)foo(/v: Unmatched ')'
CONSOLE ERROR: Pattern attribute value )foo( is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /)foo(/v: Unmatched ')'
CONSOLE ERROR: Pattern attribute value )foo( is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /)foo(/v: Unmatched ')'
CONSOLE ERROR: Pattern attribute value foo\ is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /foo\/v: \ at end of pattern
CONSOLE ERROR: Pattern attribute value [0-9 is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /[0-9/v: Unterminated character class
CONSOLE ERROR: Pattern attribute value [0-9 is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /[0-9/v: Unterminated character class
CONSOLE ERROR: Pattern attribute value \@ is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /\@/v: Invalid escape
CONSOLE ERROR: Pattern attribute value \p is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /\p/v: Invalid property name
CONSOLE ERROR: Pattern attribute value [(] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'patternMismatch' property from 'ValidityState': Invalid regular expression: /[(]/v: Invalid character in character class
This test checks validity.patternMismatch.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS patternMismatchFor("simple") is false
PASS patternMismatchFor("no-pattern-and-no-value") is false
PASS patternMismatchFor("ip-address") is false
PASS patternMismatchFor("email-address") is false
PASS patternMismatchFor("wrong-email-address") is true
PASS patternMismatchFor("match-1") is false
PASS patternMismatchFor("match-2") is false
PASS patternMismatchFor("match-3") is false
PASS patternMismatchFor("match-4") is false
PASS patternMismatchFor("match-5") is false
PASS patternMismatchFor("match-6") is false
PASS patternMismatchFor("match-7") is false
PASS patternMismatchFor("match-8") is false
PASS patternMismatchFor("match-9") is false
PASS patternMismatchFor("match-10") is false
PASS patternMismatchFor("match-11") is false
PASS patternMismatchFor("match-12") is false
PASS patternMismatchFor("match-13") is false
PASS patternMismatchFor("match-14") is false
PASS patternMismatchFor("match-15") is false
PASS patternMismatchFor("match-16") is false
PASS patternMismatchFor("match-17") is false
PASS patternMismatchFor("match-18") is false
PASS patternMismatchFor("match-19") is false
PASS patternMismatchFor("match-20") is false
PASS patternMismatchFor("wrong-gray-or-grey") is true
PASS patternMismatchFor("gray") is false
PASS patternMismatchFor("grey") is false
PASS patternMismatchFor("empty-gray-or-grey") is false
PASS patternMismatchFor("mismatch-1") is true
PASS patternMismatchFor("mismatch-2") is true
PASS patternMismatchFor("mismatch-3") is true
PASS patternMismatchFor("mismatch-4") is true
PASS patternMismatchFor("mismatch-5") is true
PASS patternMismatchFor("mismatch-6") is true
PASS patternMismatchFor("mismatch-7") is true
PASS patternMismatchFor("mismatch-8") is true
PASS patternMismatchFor("mismatch-9") is true
PASS patternMismatchFor("mismatch-10") is true
PASS patternMismatchFor("mismatch-11") is true
PASS patternMismatchFor("mismatch-12") is true
PASS patternMismatchFor("mismatch-13") is true
PASS patternMismatchFor("mismatch-14") is true
PASS patternMismatchFor("mismatch-15") is true
PASS patternMismatchFor("mismatch-16") is true
PASS patternMismatchFor("mismatch-17") is true
PASS patternMismatchFor("mismatch-18") is true
PASS patternMismatchFor("mismatch-19") is true
PASS patternMismatchFor("mismatch-20") is true
PASS patternMismatchFor("mismatch-21") is true
PASS patternMismatchFor("empty-pattern-match") is false
PASS patternMismatchFor("empty-pattern-mismatch") is true
PASS patternMismatchFor("invalid-01") is false
PASS patternMismatchFor("invalid-02") is false
PASS patternMismatchFor("invalid-03") is false
PASS patternMismatchFor("invalid-04") is false
PASS patternMismatchFor("invalid-05") is false
PASS patternMismatchFor("invalid-06") is false
PASS patternMismatchFor("disabled") is true
PASS patternMismatchFor("unicode-invalid-01") is false
PASS patternMismatchFor("unicode-invalid-02") is false
PASS patternMismatchFor("unicode-invalid-02") is false
PASS patternMismatchFor("unicode-invalid-03") is false
PASS patternMismatchFor("unicode-match-01") is false
PASS patternMismatchFor("unicode-match-02") is false
PASS patternMismatchFor("unicode-match-03") is false
PASS patternMismatchFor("unicode-match-04") is false
PASS patternMismatchFor("unicode-mismatch-01") is true
PASS successfullyParsed is true

TEST COMPLETE
